% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logistica_estudio_cohorte}
\alias{logistica_estudio_cohorte}
\title{Cálculo Logístico para Estudios de Cohorte}
\usage{
logistica_estudio_cohorte(
  n_expuestos,
  n_no_expuestos,
  tasa_reclutamiento,
  tasa_rechazo,
  tasa_perdida_seguimiento,
  tasa_elegibilidad,
  dias_laborables_mes
)
}
\arguments{
\item{n_expuestos}{Número de sujetos expuestos requeridos}

\item{n_no_expuestos}{Número de sujetos no expuestos requeridos}

\item{tasa_reclutamiento}{Tasa de reclutamiento por día}

\item{tasa_rechazo}{Tasa de rechazo esperada}

\item{tasa_perdida_seguimiento}{Tasa esperada de pérdida en el seguimiento}

\item{tasa_elegibilidad}{Tasa de elegibilidad esperada}

\item{dias_laborables_mes}{Número de días laborables por mes}
}
\value{
Una lista con cálculos logísticos para el estudio
}
\description{
Cálculo Logístico para Estudios de Cohorte
}
