% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SampleRiskRatioMulti}
\alias{SampleRiskRatioMulti}
\title{Cálculo de Tamaño Muestral y Potencia para Estudios de Cohorte}
\usage{
SampleRiskRatioMulti(
  alfa,
  potencia = NULL,
  n_expuestos = NULL,
  p0 = NULL,
  RR = NULL,
  p1 = NULL,
  IC_superior = NULL,
  IC_inferior = NULL,
  EE = NULL,
  n_previo = NULL,
  r = 1,
  metodo = "rho",
  valores_rho = seq(0, 0.9, by = 0.1)
)
}
\arguments{
\item{alfa}{Nivel de significación}

\item{potencia}{Potencia deseada (opcional si se proporciona n_expuestos)}

\item{n_expuestos}{Número de sujetos expuestos (opcional si se proporciona potencia)}

\item{p0}{Incidencia en grupo no expuesto (requerido si metodo = "rho")}

\item{RR}{Riesgo Relativo a detectar (opcional si se proporciona p1)}

\item{p1}{Incidencia en grupo expuesto (opcional si se proporciona RR)}

\item{IC_superior}{Límite superior del intervalo de confianza del RR}

\item{IC_inferior}{Límite inferior del intervalo de confianza del RR (opcional si se proporciona EE)}

\item{EE}{Error estándar del coeficiente (opcional si se proporcionan los IC)}

\item{n_previo}{Tamaño de muestra del estudio previo (requerido para método "ee")}

\item{r}{Razón de no expuestos a expuestos (n_no_expuestos / n_expuestos)}

\item{metodo}{Método de cálculo: "rho" o "ee" (error estándar)}

\item{valores_rho}{Vector de valores rho para calcular tamaños muestrales (si metodo = "rho")}
}
\value{
Un data frame con resultados según el método elegido
}
\description{
Cálculo de Tamaño Muestral y Potencia para Estudios de Cohorte
}
